FIXED: Now Personal Game Page charge from db images

+++ mysql/CreateDB.sql
@@ -7,7 +7,7 @@ USE GameShop;
 
 CREATE TABLE Game (
 	id int NOT NULL AUTO_INCREMENT,
-    price float NOT NULL,
+    price int NOT NULL,
     name varchar(30) NOT NULL,
     description text DEFAULT(""),
     shortDescription text DEFAULT(""),
